requirements.txt
alembic==1.6.2
autopep8==1.5.7
click==8.0.0
dnspython==2.1.0
email-validator==1.1.2
Flask==2.0.0
Flask-Admin==1.5.8
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-MySQL==1.5.2
Flask-PyMongo==2.3.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.5.1
Flask-Reuploaded
Flask-WTF==0.14.3
gunicorn==20.1.0
idna==3.1
is-safe-url==1.0
itsdangerous==2.0.0
Jinja2==3.0.0
Mako==1.1.4
MarkupSafe==2.0.0
mysql-connector-python==8.0.25
pip==21.1.1
protobuf==3.17.0
pycodestyle==2.7.0
pymongo==3.11.4
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.37.2
uuid
toml==0.10.2
WTForms==2.3.3
Werkzeug==2.0.0
hashids==1.3.1
